UI: fechas de Gasto como Informe de Ventas; Ajuste de Stock en tabla compacta; botones 5dp global; fix computePrecio1ByMargen; mejoras visuales y limpieza de deprecaciones

+++ app/src/main/java/ar/com/nexofiscal/nexofiscalposv2/db/entity/ComprobantePromocionEntity.kt
@@ -18,6 +18,6 @@ import androidx.room.Index
     indices = [Index(value = ["comprobanteLocalId"])]
 )
 data class ComprobantePromocionEntity(
-    val comprobanteLocalId: Long,
+    val comprobanteLocalId: Int,
     val promocionId: Int
 )
