+++ app/src/main/java/ar/com/nexofiscal/nexofiscalposv2/db/viewmodel/ComprobanteViewModel.kt
@@ -134,7 +134,7 @@ class ComprobanteViewModel(application: Application) : AndroidViewModel(applicat
             if (promociones.isNotEmpty()) {
                 val promocionEntities = promociones.map { promocion ->
                     ComprobantePromocionEntity(
-                        comprobanteLocalId = newComprobanteId,
+                        comprobanteLocalId = newComprobanteId.toInt(),
                         promocionId = promocion.id
                     )
                 }
